+++ gateway/src/router.rs
@@ -8,6 +8,7 @@ use crossbeam_channel::select;
 
 use uuid::Uuid;
 
+use crate::client_events::ClientEvents;
 use crate::logging::{short_uuid, EMPTY_SHORT_UUID};
 use crate::model::*;
